requirements.txt
requests
beautifulsoup4
yfinance
openai
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv